# docs/oxford-gmwo-swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Global Model Workstation Online (GMWO) API",
...
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://model.oxfordeconomics.com/api"
    }
  ],
  "paths": {
...
    "/v1/operations/{id}": {
      "get": {
        "tags": [
          "ForecastOperation"
        ],
        "summary": "Gets an existing forecast operation.",
        "operationId": "ForecastOperation_Get",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The id of the operation.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastOperation"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/operations/solve": {
      "post": {
        "tags": [
          "ForecastOperation"
        ],
        "summary": "Enqueues an operation which applies the specified commands or data imports to the input forecast before performing a model solution, creating a new forecast.",
        "description": "This request enqueues an operation which runs in the background. Use the /operations/{id}/await endpoint to wait for the operation to finish. It is not necessary to submit any Commands or a Commands File in order to perform a solve but, if you do, you should submit either one or the other and not both.",
        "operationId": "ForecastOperation_Solve",
        "requestBody": {
          "x-name": "solveRequest",
          "description": "The operation parameters.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolveForecast"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "The enqueued operation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastOperation"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PeriodRange": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "From",
          "To"
        ],
        "properties": {
          "From": {
            "type": "string",
            "minLength": 1
          },
          "To": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
      "ForecastOperation": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ForecastOperationHistory"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "Artifacts",
              "Resources"
            ],
            "properties": {
              "Artifacts": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ForecastArtifact"
                }
              },
              "Resources": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/OperationResource"
                }
              }
            }
          }
        ]
      },
      "ForecastArtifact": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "Id",
          "Filename",
          "Type",
          "DownloadUrl"
        ],
        "properties": {
          "Id": {
            "type": "string"
          },
          "Filename": {
            "type": "string"
          },
          "Type": {
            "type": "string"
          },
          "DownloadUrl": {
            "type": "string"
          }
        }
      },
      "OperationResource": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "Id",
          "Version",
          "Role"
        ],
        "properties": {
          "Id": {
            "type": "string"
          },
          "Path": {
            "type": "string",
            "nullable": true
          },
          "Version": {
            "type": "integer",
            "format": "int32"
          },
          "Role": {
            "$ref": "#/components/schemas/OperationResourceRole"
          }
        }
      },
      "OperationResourceRole": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Input",
          "Output"
        ],
        "enum": [
          "Input",
          "Output"
        ]
      },
      "ForecastOperationHistory": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "Id",
          "CreatedAt",
          "Status"
        ],
        "properties": {
          "Id": {
            "type": "string"
          },
          "CreatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "StartedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "CompletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "Status": {
            "$ref": "#/components/schemas/OperationStatus"
          },
          "Duration": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "FailureReason": {
            "type": "string",
            "nullable": true
          },
          "Name": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "OperationStatus": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Queued",
          "InProgress",
          "Failed",
          "Succeeded",
          "Cancelled"
        ],
        "enum": [
          "Queued",
          "InProgress",
          "Failed",
          "Succeeded",
          "Cancelled"
        ]
      },
...
      "ApplyForecastBase": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "InputForecast",
          "OutputForecast"
        ],
        "properties": {
          "InputForecast": {
            "type": "string",
            "minLength": 1
          },
          "OutputForecast": {
            "type": "string",
            "minLength": 1
          },
          "OperationName": {
            "type": "string",
            "nullable": true
          },
          "Commands": {
            "type": "string",
            "nullable": true
          },
          "CommandsFile": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "SolveForecast": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ApplyForecastBase"
          },
          {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "SolutionRange"
            ],
            "properties": {
              "SolutionRange": {
                "$ref": "#/components/schemas/PeriodRange"
              },
              "Groups": {
                "type": "array",
                "nullable": true,
                "items": {
                  "type": "string"
                }
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {
      "Bearer authentication": {
        "type": "http",
        "description": "Enter your JWT here",
        "name": "Authorization",
        "in": "header",
        "scheme": "Bearer",
        "flows": {
          "implicit": {
            "tokenUrl": "/api/v1/swagger/get-token"
          }
        }
      }
    }
  },
  "security": [
    {
      "Bearer authentication": []
    }
  ]
}